separated logic from ui and added missing translations
