+++ src/app/components/home/resume/components/hooks/cover-letter-generator/useGenerate.ts
@@ -5,6 +5,7 @@ export default function useGenerate() {
   const [isGenerating, setIsGenerating] = useState(false);
 
   const handleGenerate = (file: string) => {
+    console.log(file);
     setIsGenerating(true);
     // Simulate generation
     setTimeout(() => {
