+++ src/app/components/home/resume/components/pages/ats-scanner/result/result.tsx
@@ -1,6 +1,6 @@
 import { TextButton } from "@/app/ui-library/components/buttons/text-button";
 import React, { Dispatch, SetStateAction } from "react";
-import { Resume, ScanResults } from "../../../../types/interfaces";
+import { ScanResults } from "../../../../types/interfaces";
 import JobMatchSection from "./job-match-section";
 import SkillAccuracySection from "./skill-accuracy-section";
 import SkillApplicationSection from "./skill-application-section";
@@ -15,11 +15,11 @@ export default function Result({
   results: ScanResults;
   setResults: Dispatch<SetStateAction<ScanResults | null>>;
 }) {
-  const getFileName = (file: Resume | File | null): string => {
-    if (!file) return "";
-    if (file instanceof File) return file.name;
-    return file.name;
-  };
+  // const getFileName = (file: Resume | File | null): string => {
+  //   if (!file) return "";
+  //   if (file instanceof File) return file.name;
+  //   return file.name;
+  // };
 
   const { t } = useTranslation();
